Replaced BufferedInputStream with RandomAccessFile
In case of an larger DBF file, BufferedInputStream can cause trouble,
because of the limit of buffered bytes.
To be specific, in previous version, the method
net.iryndin.jdbf.reader.DbfReader.readMetadata() used
BufferedInputStream.mark(int) to allocate 1024*1024 byte buffer. But
then, in case the file size goes over 1MB (stream is read over 1MB)
BufferedInputStream.reset() is likely to cause IOException as the mark
is invalidated.
RandomAccessFile would not be fast as BufferedInputStream(memory buffer)
but fast enough, and is reliable on larger files.